+++ lib/device_geometry.dart
@@ -1,7 +1,9 @@
 import 'dart:io';
 
 import 'package:device_geometry/platforms/ios.dart';
+import 'package:device_geometry/widgets/provider.dart';
 import 'package:device_info_plus/device_info_plus.dart';
+import 'package:flutter/widgets.dart';
 
 class DisplayCornerRadii {
   final double topLeft;
@@ -24,14 +26,32 @@ class DisplayGeometry {
 }
 
 class DeviceGeometry {
+  static Future<DeviceGeometry?>? _future;
   final DisplayGeometry display;
 
   const DeviceGeometry({required this.display});
 
-  Future<DeviceGeometry?> get() async {
+  static Future<DeviceGeometry?> get() async {
+    return _future ??= _getImpl();
+  }
+
+  static Future<DeviceGeometry?> _getImpl() async {
     if (Platform.isIOS) {
       return (await DeviceInfoPlugin().iosInfo).geometry;
     }
     return null;
+  }
+
+  static DeviceGeometry? maybeOf(BuildContext context) {
+    return context
+        .dependOnInheritedWidgetOfExactType<DeviceGeometryProvider>()
+        ?.geometry;
+  }
+
+  static DeviceGeometry? of(BuildContext context) {
+    final DeviceGeometryProvider? provider = context
+        .dependOnInheritedWidgetOfExactType<DeviceGeometryProvider>();
+    assert(provider != null, 'No DeviceGeometryProvider found');
+    return provider!.geometry;
   }
 }