+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Add DeviceGeometryProvider provider
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial implementation for corner radii on iOS